refactor(database/mod.rs): Moved contents to queries/lib.rs and removed the database module from bld_core
