+++ crates/bld_entities/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0.40"
+bld_config = { path = "../bld_config" }
 bld_migrations = { path = "../bld_migrations" }
 chrono = { version = "0.4.23", default-features = false, features = ["std"] }
 sea-orm = { version = "0.12.2", features = ["sqlx-postgres", "sqlx-mysql", "sqlx-sqlite", "runtime-tokio-rustls"] }
